+++ server/controllers/api.js
@@ -11,12 +11,13 @@ const chalk = require("chalk"); // https://www.npmjs.com/package/chalk
 const moment = require("moment"); // https://momentjs.com/
 const mongoose = require("mongoose"); // http://mongoosejs.com/
 const Game = mongoose.model("Game");
+const Planet = mongoose.model("Planet");
 const logger = require("../handlers/logger/console");
 const cfg = require("../config");
 
 
 /*
- * prepare an array of text strings from filesystem
+ * prepare an array of text strings from filesystem ====================================================================
  * this array is kept in memory and can be accessed fast by the api endpoint
  */
 const allTexts = [];
@@ -64,7 +65,7 @@ locales.forEach(locale => {
 logger.info(`[App] cached ${chalk.yellow(allTexts.length)} client text objects.`);
 
 /*
- * get game status
+ * get game status =====================================================================================================
  *
  * @param {ExpressHTTPRequest} req
  * @param {ExpressHTTPResponse} res
@@ -101,7 +102,7 @@ exports.gameStatus = async (req, res, next) => {
 };
 
 /*
- * get text strings
+ * get text strings ====================================================================================================
  *
  * @param {ExpressHTTPRequest} req
  * @param {ExpressHTTPResponse} res
@@ -116,4 +117,80 @@ exports.getTextStrings = async (req, res) => {
     } else {
         res.json({error: "could not get client text strings."});
     }
+};
+
+/*
+ * check if the current user is allowed to get game data ===============================================================
+ * @param {ExpressHTTPRequest} req
+ * @param {ExpressHTTPResponse} res
+ * @param {callback} next
+ */
+exports.checkGameDataPermission = (req, res, next) => {
+    let failed = false;
+    if (`${req.user.selectedPlayer._id}` !== req.params.playerid) {
+        logger.error("spoofing attempt for game data via playerid.");
+        failed = true;
+    }
+    if (`${req.user.selectedPlayer.game._id}` !== req.params.gameid) {
+        logger.error("spoofing attempt for game data via gameid.");
+        failed = true;
+    }
+    if (failed) {
+        return res.json({});
+    } else {
+        next();
+    }
+};
+
+/*
+ * get game data =======================================================================================================
+ * we do not want to send full mongoose objects to the client,
+ * (there might be parts that we don't want to tell the player)
+ * so we massage the data a bit.
+ * @param {ExpressHTTPRequest} req
+ * @param {ExpressHTTPResponse} res
+ */
+exports.getGameData = async (req, res) => {
+    const player = req.user.selectedPlayer;
+    const game = player.game;
+    const stars = player.stars.map(star => star.id);
+    const planets = await Planet.find({star: {$in: stars}});
+    const returnData = {
+        game: {
+            number: game.number,
+            active: game.active,
+            dimensions: game.dimensions,
+            turn: game.turn,
+            turnDue: game.turnDue,
+            turnDuration: game.turnDuration,
+            processing: game.processing
+        },
+        player: {
+            name: player.name,
+            ticker: player.ticker
+        },
+        stars: [],
+    };
+    player.stars.length && player.stars.forEach( star => {
+        let starPlanets = [];
+        planets.length && planets.forEach( planet => {
+            // make sure the planet belongs to this star before adding
+            if (`${planet.star}` === `${star._id}`) {
+                starPlanets.push({
+                    id: planet._id,
+                    type: planet.type,
+                    orbitalIndex: planet.orbitalIndex
+                });
+            }
+        });
+        returnData.stars.push({
+            id: star._id,
+            name: star.name,
+            spectral: star.spectral,
+            coordX: star.coordX,
+            coordY: star.coordY,
+            planets: starPlanets
+        });
+    });
+    return res.json(returnData);
 };